# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Robot Rescue API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type IncidentIssueType =
  (typeof IncidentIssueType)[keyof typeof IncidentIssueType];

export const IncidentIssueType = {
  obstacle_detected: "obstacle_detected",
  system_error: "system_error",
  path_blocked: "path_blocked",
  sensor_failure: "sensor_failure",
  battery_critical: "battery_critical",
  communication_lost: "communication_lost",
} as const;

export type IncidentSeverity =
  (typeof IncidentSeverity)[keyof typeof IncidentSeverity];

export const IncidentSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type IncidentStatus =
  (typeof IncidentStatus)[keyof typeof IncidentStatus];

export const IncidentStatus = {
  waiting: "waiting",
  in_progress: "in_progress",
  resolved: "resolved",
} as const;

export interface SensorData {
  battery: number;
  speed: number;
  temperature: number;
  obstacleDistance?: number | null;
  signalStrength: number;
}

export interface Incident {
  id: string;
  robotId: string;
  location: string;
  issueType: IncidentIssueType;
  severity: IncidentSeverity;
  status: IncidentStatus;
  description: string;
  timestamp: string;
  resolvedAt?: string | null;
  actionTaken?: string | null;
  responseTimeSeconds?: number | null;
  sensorData: SensorData;
}

export type CreateIncidentBodyIssueType =
  (typeof CreateIncidentBodyIssueType)[keyof typeof CreateIncidentBodyIssueType];

export const CreateIncidentBodyIssueType = {
  obstacle_detected: "obstacle_detected",
  system_error: "system_error",
  path_blocked: "path_blocked",
  sensor_failure: "sensor_failure",
  battery_critical: "battery_critical",
  communication_lost: "communication_lost",
} as const;

export type CreateIncidentBodySeverity =
  (typeof CreateIncidentBodySeverity)[keyof typeof CreateIncidentBodySeverity];

export const CreateIncidentBodySeverity = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateIncidentBody {
  robotId: string;
  location: string;
  issueType: CreateIncidentBodyIssueType;
  severity: CreateIncidentBodySeverity;
  description: string;
}

export type UpdateIncidentBodyStatus =
  (typeof UpdateIncidentBodyStatus)[keyof typeof UpdateIncidentBodyStatus];

export const UpdateIncidentBodyStatus = {
  waiting: "waiting",
  in_progress: "in_progress",
  resolved: "resolved",
} as const;

export type UpdateIncidentBodyActionTaken =
  (typeof UpdateIncidentBodyActionTaken)[keyof typeof UpdateIncidentBodyActionTaken];

export const UpdateIncidentBodyActionTaken = {
  reroute: "reroute",
  pause: "pause",
  manual_override: "manual_override",
  escalate: "escalate",
} as const;

export interface UpdateIncidentBody {
  status?: UpdateIncidentBodyStatus;
  actionTaken?: UpdateIncidentBodyActionTaken;
}

export interface IncidentLogEntry {
  id: string;
  robotId: string;
  issueType: string;
  actionTaken?: string | null;
  resolvedAt?: string | null;
  responseTimeSeconds?: number | null;
  severity: string;
  location: string;
}

export type AnalyticsSummaryIssueBreakdownItem = {
  issueType: string;
  count: number;
};

export type AnalyticsSummarySeverityBreakdownItem = {
  severity: string;
  count: number;
};

export interface AnalyticsSummary {
  totalIncidents: number;
  activeIncidents: number;
  resolvedIncidents: number;
  avgResponseTimeSeconds: number;
  issueBreakdown: AnalyticsSummaryIssueBreakdownItem[];
  severityBreakdown: AnalyticsSummarySeverityBreakdownItem[];
  recentActivity: Incident[];
}

export type ListIncidentsParams = {
  status?: ListIncidentsStatus;
};

export type ListIncidentsStatus =
  (typeof ListIncidentsStatus)[keyof typeof ListIncidentsStatus];

export const ListIncidentsStatus = {
  active: "active",
  resolved: "resolved",
  all: "all",
} as const;