# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Robot Rescue API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all incidents
 */
export const ListIncidentsQueryParams = zod.object({
  status: zod.enum(["active", "resolved", "all"]).optional(),
});

export const ListIncidentsResponseItem = zod.object({
  id: zod.string(),
  robotId: zod.string(),
  location: zod.string(),
  issueType: zod.enum([
    "obstacle_detected",
    "system_error",
    "path_blocked",
    "sensor_failure",
    "battery_critical",
    "communication_lost",
  ]),
  severity: zod.enum(["low", "medium", "high"]),
  status: zod.enum(["waiting", "in_progress", "resolved"]),
  description: zod.string(),
  timestamp: zod.string(),
  resolvedAt: zod.string().nullish(),
  actionTaken: zod.string().nullish(),
  responseTimeSeconds: zod.number().nullish(),
  sensorData: zod.object({
    battery: zod.number(),
    speed: zod.number(),
    temperature: zod.number(),
    obstacleDistance: zod.number().nullish(),
    signalStrength: zod.number(),
  }),
});
export const ListIncidentsResponse = zod.array(ListIncidentsResponseItem);

/**
 * @summary Create a new incident
 */
export const CreateIncidentBody = zod.object({
  robotId: zod.string(),
  location: zod.string(),
  issueType: zod.enum([
    "obstacle_detected",
    "system_error",
    "path_blocked",
    "sensor_failure",
    "battery_critical",
    "communication_lost",
  ]),
  severity: zod.enum(["low", "medium", "high"]),
  description: zod.string(),
});

/**
 * @summary Get an incident by ID
 */
export const GetIncidentParams = zod.object({
  id: zod.coerce.string(),
});

export const GetIncidentResponse = zod.object({
  id: zod.string(),
  robotId: zod.string(),
  location: zod.string(),
  issueType: zod.enum([
    "obstacle_detected",
    "system_error",
    "path_blocked",
    "sensor_failure",
    "battery_critical",
    "communication_lost",
  ]),
  severity: zod.enum(["low", "medium", "high"]),
  status: zod.enum(["waiting", "in_progress", "resolved"]),
  description: zod.string(),
  timestamp: zod.string(),
  resolvedAt: zod.string().nullish(),
  actionTaken: zod.string().nullish(),
  responseTimeSeconds: zod.number().nullish(),
  sensorData: zod.object({
    battery: zod.number(),
    speed: zod.number(),
    temperature: zod.number(),
    obstacleDistance: zod.number().nullish(),
    signalStrength: zod.number(),
  }),
});

/**
 * @summary Update incident status or take action
 */
export const UpdateIncidentParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateIncidentBody = zod.object({
  status: zod.enum(["waiting", "in_progress", "resolved"]).optional(),
  actionTaken: zod
    .enum(["reroute", "pause", "manual_override", "escalate"])
    .optional(),
});

export const UpdateIncidentResponse = zod.object({
  id: zod.string(),
  robotId: zod.string(),
  location: zod.string(),
  issueType: zod.enum([
    "obstacle_detected",
    "system_error",
    "path_blocked",
    "sensor_failure",
    "battery_critical",
    "communication_lost",
  ]),
  severity: zod.enum(["low", "medium", "high"]),
  status: zod.enum(["waiting", "in_progress", "resolved"]),
  description: zod.string(),
  timestamp: zod.string(),
  resolvedAt: zod.string().nullish(),
  actionTaken: zod.string().nullish(),
  responseTimeSeconds: zod.number().nullish(),
  sensorData: zod.object({
    battery: zod.number(),
    speed: zod.number(),
    temperature: zod.number(),
    obstacleDistance: zod.number().nullish(),
    signalStrength: zod.number(),
  }),
});

/**
 * @summary Get resolved/logged incidents
 */
export const GetIncidentLogResponseItem = zod.object({
  id: zod.string(),
  robotId: zod.string(),
  issueType: zod.string(),
  actionTaken: zod.string().nullish(),
  resolvedAt: zod.string().nullish(),
  responseTimeSeconds: zod.number().nullish(),
  severity: zod.string(),
  location: zod.string(),
});
export const GetIncidentLogResponse = zod.array(GetIncidentLogResponseItem);

/**
 * @summary Get analytics summary
 */
export const GetAnalyticsSummaryResponse = zod.object({
  totalIncidents: zod.number(),
  activeIncidents: zod.number(),
  resolvedIncidents: zod.number(),
  avgResponseTimeSeconds: zod.number(),
  issueBreakdown: zod.array(
    zod.object({
      issueType: zod.string(),
      count: zod.number(),
    }),
  ),
  severityBreakdown: zod.array(
    zod.object({
      severity: zod.string(),
      count: zod.number(),
    }),
  ),
  recentActivity: zod.array(
    zod.object({
      id: zod.string(),
      robotId: zod.string(),
      location: zod.string(),
      issueType: zod.enum([
        "obstacle_detected",
        "system_error",
        "path_blocked",
        "sensor_failure",
        "battery_critical",
        "communication_lost",
      ]),
      severity: zod.enum(["low", "medium", "high"]),
      status: zod.enum(["waiting", "in_progress", "resolved"]),
      description: zod.string(),
      timestamp: zod.string(),
      resolvedAt: zod.string().nullish(),
      actionTaken: zod.string().nullish(),
      responseTimeSeconds: zod.number().nullish(),
      sensorData: zod.object({
        battery: zod.number(),
        speed: zod.number(),
        temperature: zod.number(),
        obstacleDistance: zod.number().nullish(),
        signalStrength: zod.number(),
      }),
    }),
  ),
});