# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Robot Rescue API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyticsSummaryIssueBreakdownItem } from "./analyticsSummaryIssueBreakdownItem";
import type { AnalyticsSummarySeverityBreakdownItem } from "./analyticsSummarySeverityBreakdownItem";
import type { Incident } from "./incident";

export interface AnalyticsSummary {
  totalIncidents: number;
  activeIncidents: number;
  resolvedIncidents: number;
  avgResponseTimeSeconds: number;
  issueBreakdown: AnalyticsSummaryIssueBreakdownItem[];
  severityBreakdown: AnalyticsSummarySeverityBreakdownItem[];
  recentActivity: Incident[];
}