# lib/api-zod/src/generated/types/createIncidentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Robot Rescue API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateIncidentBodyIssueType } from "./createIncidentBodyIssueType";
import type { CreateIncidentBodySeverity } from "./createIncidentBodySeverity";

export interface CreateIncidentBody {
  robotId: string;
  location: string;
  issueType: CreateIncidentBodyIssueType;
  severity: CreateIncidentBodySeverity;
  description: string;
}