# lib/api-zod/src/generated/types/incident.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Robot Rescue API
 * OpenAPI spec version: 0.1.0
 */
import type { IncidentIssueType } from "./incidentIssueType";
import type { IncidentSeverity } from "./incidentSeverity";
import type { IncidentStatus } from "./incidentStatus";
import type { SensorData } from "./sensorData";

export interface Incident {
  id: string;
  robotId: string;
  location: string;
  issueType: IncidentIssueType;
  severity: IncidentSeverity;
  status: IncidentStatus;
  description: string;
  timestamp: string;
  resolvedAt?: string | null;
  actionTaken?: string | null;
  responseTimeSeconds?: number | null;
  sensorData: SensorData;
}