# lib/api-zod/src/generated/types/incidentLogEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Robot Rescue API
 * OpenAPI spec version: 0.1.0
 */

export interface IncidentLogEntry {
  id: string;
  robotId: string;
  issueType: string;
  actionTaken?: string | null;
  resolvedAt?: string | null;
  responseTimeSeconds?: number | null;
  severity: string;
  location: string;
}