# lib/api-zod/src/generated/types/sensorData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Robot Rescue API
 * OpenAPI spec version: 0.1.0
 */

export interface SensorData {
  battery: number;
  speed: number;
  temperature: number;
  obstacleDistance?: number | null;
  signalStrength: number;
}