# lib/api-zod/src/generated/types/updateIncidentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Robot Rescue API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateIncidentBodyActionTaken } from "./updateIncidentBodyActionTaken";
import type { UpdateIncidentBodyStatus } from "./updateIncidentBodyStatus";

export interface UpdateIncidentBody {
  status?: UpdateIncidentBodyStatus;
  actionTaken?: UpdateIncidentBodyActionTaken;
}